+++ CHANGELOG.md
@@ -34,6 +34,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - `client.batch.hisDelete(ids, range, options)`
   - `client.batch.create(entities, options)`
   - `client.batch.update(entities, options)`
+  - `client.batch.deleteById(ids, options)`
 
 ### CHANGED
 - Client no longer throws a Axios error if a response has been received and response is a Haystack of GraphQL error.